+++ finance_utils.py
@@ -14,7 +14,6 @@ import time
 
 # Custom
 import pandas as pd
-import numpy as np
 
 # User
 import yqd
@@ -201,53 +200,3 @@ def validate_symbol_data(csv_file):
         valid = False
     f.close()
     return valid
-
-
-def _main():
-    sf = 'stock_db/dj.txt'
-    print("symbol file {} contains the following stocks: {}".format(sf, get_symbols_from_file(sf)))
-
-    d = './stock_db/test'
-
-    s = 'SPY'
-    f = symbol_to_filename(s, d)
-    print("symbol {} with directory {} gives filename {}".format(s, d, f))
-    print("filename {} gives symbol {}".format(f, filename_to_symbol(f)))
-    print("filename {} gives symbol {}".format(f.upper(), filename_to_symbol(f.upper())))
-    print("validate_symbol_data {} = {}".format(f, validate_symbol_data(f)))
-
-    print("directory {} contains the following stocks: {}".format(d, get_all_symbols(d)))
-
-    start_date = datetime.date(1900, 1, 1)
-    end_date = datetime.date.today()
-
-    if False:
-        download_data(s, d, start_date, end_date)
-        update_all_symbols(d, start_date, end_date)
-    df = load_data_frame(f, datetime.date(2018, 1, 1), datetime.date(2018, 4, 1))
-    print(df.describe())
-    print(df.head())
-
-    print(get_date(df)[0:3])
-    print(get_open(df)[0:3])
-    print(get_high(df)[0:3])
-    print(get_low(df)[0:3])
-    print(get_close(df)[0:3])
-    print(get_volume(df)[0:3])
-
-    # Not applicable for a single stock, but just to test...
-    print(normalize_data_frame(df).head())
-
-    # Test by adding some NaN
-    df.iloc[0:10, 0] = np.nan  # beginning
-    df.iloc[11:20, 1] = np.nan  # middle
-    df.iloc[-10:, 2] = np.nan  # end
-    print(df.isna().any())
-    fill_nan_data(df)
-    print(df.isna().any())
-
-    print(download_url("https://www.google.ca")[0:100])
-
-
-if __name__ == '__main__':
-    _main()
